backend/scripts/import_data.py
#!/usr/bin/env python
"""
Скрипт для импорта данных из старой базы данных MySQL в новую PostgreSQL.
Использует прямое соединение с обеими базами и выполняет миграцию данных.

Запуск:
python -m scripts.import_data

Требуется установить дополнительные зависимости:
pip install mysql-connector-python
"""

import logging
import os
import sys
from datetime import datetime

# Добавляем корневую директорию проекта в PYTHONPATH
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

# Импорты из приложения
from sqlalchemy.orm import Session
from app.db.session import SessionLocal
# from app.models.user import User
# from app.models.product import Product
# from app.models.category import Category
# from app.models.order import Order

# Настройка логирования
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    handlers=[
        logging.FileHandler(f"import_data_{datetime.now().strftime('%Y%m%d_%H%M%S')}.log"),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

# Конфигурация подключения к старой базе данных MySQL
MYSQL_CONFIG = {
    "host": "localhost",
    "user": "root",
    "password": "",
    "database": "mrauto"
}


def import_users(mysql_cursor, db: Session):
    """Импорт пользователей"""
    logger.info("Начинаем импорт пользователей")
    # TODO: Реализовать импорт пользователей
    # mysql_cursor.execute("SELECT * FROM users")
    # users = mysql_cursor.fetchall()
    # for user in users:
    #     db_user = User(
    #         id=user["id"],
    #         email=user["email"],
    #         hashed_password=user["password"],  # Внимание: необходима обработка хешей!
    #         full_name=user["full_name"],
    #         is_active=True,
    #         is_superuser=user["is_admin"] == 1
    #     )
    #     db.add(db_user)
    # db.commit()
    # logger.info(f"Импортировано {len(users)} пользователей")


def import_categories(mysql_cursor, db: Session):
    """Импорт категорий товаров"""
    logger.info("Начинаем импорт категорий")
    # TODO: Реализовать импорт категорий
    # mysql_cursor.execute("SELECT * FROM categories")
    # categories = mysql_cursor.fetchall()
    # for category in categories:
    #     db_category = Category(
    #         id=category["id"],
    #         name=category["name"],
    #         description=category["description"],
    #         slug=category["slug"]
    #     )
    #     db.add(db_category)
    # db.commit()
    # logger.info(f"Импортировано {len(categories)} категорий")


def import_products(mysql_cursor, db: Session):
    """Импорт товаров"""
    logger.info("Начинаем импорт товаров")
    # TODO: Реализовать импорт товаров
    # mysql_cursor.execute("SELECT * FROM products")
    # products = mysql_cursor.fetchall()
    # for product in products:
    #     db_product = Product(
    #         id=product["id"],
    #         name=product["name"],
    #         description=product["description"],
    #         price=product["price"],
    #         category_id=product["category_id"],
    #         sku=product["sku"],
    #         stock=product["stock"],
    #         is_active=product["is_active"] == 1
    #     )
    #     db.add(db_product)
    # db.commit()
    # logger.info(f"Импортировано {len(products)} товаров")


def import_orders(mysql_cursor, db: Session):
    """Импорт заказов"""
    logger.info("Начинаем импорт заказов")
    # TODO: Реализовать импорт заказов
    # mysql_cursor.execute("SELECT * FROM orders")
    # orders = mysql_cursor.fetchall()
    # for order in orders:
    #     db_order = Order(
    #         id=order["id"],
    #         user_id=order["user_id"],
    #         total_amount=order["total_amount"],
    #         status=order["status"],
    #         created_at=order["created_at"]
    #     )
    #     db.add(db_order)
    # db.commit()
    # logger.info(f"Импортировано {len(orders)} заказов")


def main():
    logger.info("Начинаем импорт данных из старой БД")
    
    try:
        # Подключение к MySQL
        import mysql.connector
        mysql_conn = mysql.connector.connect(**MYSQL_CONFIG)
        mysql_cursor = mysql_conn.cursor(dictionary=True)
        
        # Подключение к PostgreSQL через SQLAlchemy
        db = SessionLocal()
        
        # Импорт данных
        import_users(mysql_cursor, db)
        import_categories(mysql_cursor, db)
        import_products(mysql_cursor, db)
        import_orders(mysql_cursor, db)
        
        logger.info("Импорт данных успешно завершен")
        
    except Exception as e:
        logger.error(f"Ошибка при импорте данных: {e}")
        raise
    finally:
        if 'mysql_cursor' in locals():
            mysql_cursor.close()
        if 'mysql_conn' in locals():
            mysql_conn.close()
        if 'db' in locals():
            db.close()


if __name__ == "__main__":
    main() 