backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.21
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.0
pytest==7.4.2
black==23.9.1
isort==5.12.0
flake8==6.1.0
mysql-connector-python==8.2.0 